Task-3 - Rename areEditable set to editedItems
