+++ onboarding-task/src/components/List.jsx
@@ -10,7 +10,7 @@ class List extends PureComponent {
 
   static propTypes = {
     list: ImmutablePropTypes.orderedMapOf(ImmutablePropTypes.record).isRequired,
-    areEditable: ImmutablePropTypes.setOf(PropTypes.string).isRequired,
+    editedItems: ImmutablePropTypes.setOf(PropTypes.string).isRequired,
     addItem: PropTypes.func.isRequired,
     deleteItem: PropTypes.func.isRequired,
     updateItem: PropTypes.func.isRequired,
@@ -24,7 +24,7 @@ class List extends PureComponent {
   }
 
   _getItemToRender(item, index) {
-    if (this.props.areEditable.has(item.id)) {
+    if (this.props.editedItems.has(item.id)) {
       return (<EditForm
         item={item}
         index={index}
