+++ onboarding-task/src/reducers/listApp.js
@@ -1,11 +1,11 @@
 import { combineReducers } from 'redux';
 
 import { list } from './listReducer.js';
-import { areEditable } from './areEditableReducer.js';
+import { editedItems } from './editedItemsReducer.js';
 
 const listApp = combineReducers({
   list,
-  areEditable,
+  editedItems,
 });
 
 export { listApp };
